git-svn-id: https://svn.r-project.org/R/branches/R-BIGSEXP@68552 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/extra/intl/langprefs.c
@@ -58,7 +58,7 @@ _nl_language_preferences_default (void)
 	    && CFGetTypeID (preferences) == CFArrayGetTypeID ())
 	  {
 	    CFArrayRef prefArray = (CFArrayRef)preferences;
-	    int n = CFArrayGetCount (prefArray);
+	    long n = CFArrayGetCount (prefArray);
 	    char buf[256];
 	    size_t size = 0;
 	    int i;
